requirements.txt
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-jaeger==1.21.0